FPET-967: splitted hearing controller to fix pmd issue
