+++ src/test/java/uk/gov/hmcts/reform/hmc/api/controllers/HearingControllerNextHearingDetailsTest.java
@@ -1,0 +1,205 @@
+package uk.gov.hmcts.reform.hmc.api.controllers;
+
+import static feign.Request.HttpMethod.GET;
+import static java.nio.charset.StandardCharsets.UTF_8;
+import static org.apache.commons.lang3.StringUtils.EMPTY;
+import static org.mockito.ArgumentMatchers.any;
+
+import feign.FeignException;
+import feign.Request;
+import feign.Response;
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import org.json.simple.parser.ParseException;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.MockitoAnnotations;
+import org.mockito.junit.MockitoJUnitRunner;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.test.context.ActiveProfiles;
+import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
+import uk.gov.hmcts.reform.hmc.api.model.ccd.NextHearingDetails;
+import uk.gov.hmcts.reform.hmc.api.model.response.CaseHearing;
+import uk.gov.hmcts.reform.hmc.api.model.response.HearingDaySchedule;
+import uk.gov.hmcts.reform.hmc.api.model.response.Hearings;
+import uk.gov.hmcts.reform.hmc.api.services.HearingsService;
+import uk.gov.hmcts.reform.hmc.api.services.IdamAuthService;
+import uk.gov.hmcts.reform.hmc.api.services.NextHearingDetailsService;
+
+@RunWith(MockitoJUnitRunner.class)
+@ActiveProfiles("test")
+public class HearingControllerNextHearingDetailsTest {
+
+    @InjectMocks private HearingsController hearingsController;
+
+    @Mock private IdamAuthService idamAuthService;
+
+    @Mock private AuthTokenGenerator authTokenGenerator;
+
+    @Mock private HearingsService hearingsService;
+
+    @Mock private NextHearingDetailsService nextHearingDetailsService;
+
+    private Hearings hearings;
+
+    @BeforeEach
+    void setUp() {
+
+        MockitoAnnotations.openMocks(this);
+        LocalDateTime testNextHearingDate = LocalDateTime.of(2024, 04, 28, 1, 0);
+
+        HearingDaySchedule hearingDaySchedule =
+            HearingDaySchedule.hearingDayScheduleWith()
+                .hearingVenueId("231596")
+                .hearingJudgeId("4925644")
+                .hearingStartDateTime(testNextHearingDate)
+                .build();
+        List<HearingDaySchedule> hearingDayScheduleList = new ArrayList<>();
+        hearingDayScheduleList.add(hearingDaySchedule);
+
+        CaseHearing caseHearing =
+            CaseHearing.caseHearingWith()
+                .hmcStatus("LISTED")
+                .hearingDaySchedule(hearingDayScheduleList)
+                .build();
+        List<CaseHearing> caseHearingList = new ArrayList<>();
+        caseHearingList.add(caseHearing);
+    }
+
+
+
+    @Test
+    void hearingsControllerNextHearingDateTest() throws IOException, ParseException {
+        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
+        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
+        ResponseEntity<Object> nextHearingDetailsResponse =
+            hearingsController.updateNextHearingDetails(
+                "Bearer auth", "Bearer sauth", "caseRef");
+        Assertions.assertEquals(HttpStatus.OK, nextHearingDetailsResponse.getStatusCode());
+    }
+
+    @Test
+    void nextHearingDateByCaseRefNoControllerUnauthorisedExceptionTest()
+        throws IOException, ParseException {
+
+        ResponseEntity<Object> nextHearingDetails =
+            hearingsController.updateNextHearingDetails("", "", "caseRef");
+
+        Assertions.assertEquals(HttpStatus.UNAUTHORIZED, nextHearingDetails.getStatusCode());
+    }
+
+    @Test
+    void nextHearingDateByCaseRefNoControllerFeignExceptionTest()
+        throws IOException, ParseException {
+        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
+        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
+
+        Mockito.when(nextHearingDetailsService.updateNextHearingDetails("auth", hearings))
+            .thenThrow(feignException(HttpStatus.INTERNAL_SERVER_ERROR.value(), "Not found"));
+
+        ResponseEntity<Object> nextHearingUpdateResp =
+            hearingsController.updateNextHearingDetails("auth", "sauth", "testcase");
+
+        Assertions.assertEquals(
+            HttpStatus.INTERNAL_SERVER_ERROR, nextHearingUpdateResp.getStatusCode());
+    }
+
+
+    @Test
+    void nextHearingDateByCaseRefNoInternalServerErrorTest()
+        throws IOException, ParseException {
+        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
+        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
+
+        Mockito.when(nextHearingDetailsService.updateNextHearingDetails("auth", hearings))
+            .thenThrow(new RuntimeException());
+
+        ResponseEntity<Object> nextHearingUpdateResp =
+            hearingsController.updateNextHearingDetails("auth", "sauth", "testcase");
+
+        Assertions.assertEquals(
+            HttpStatus.INTERNAL_SERVER_ERROR, nextHearingUpdateResp.getStatusCode());
+    }
+
+
+    @Test
+    void getNextHearingDateTest() throws IOException, ParseException {
+        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
+        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
+        LocalDateTime testNextHearingDate = LocalDateTime.of(2024, 04, 28, 1, 0);
+        NextHearingDetails nextHearingDetails =
+            NextHearingDetails.builder()
+                .hearingID(123L)
+                .hearingDateTime(testNextHearingDate)
+                .build();
+        Mockito.when(nextHearingDetailsService.getNextHearingDate(hearings))
+            .thenReturn(nextHearingDetails);
+        Mockito.when(authTokenGenerator.generate()).thenReturn("MOCK_S2S_TOKEN");
+        Hearings hearingsObj = Hearings.hearingsWith().caseRef("123").hmctsServiceCode("BBA3").build();
+        Mockito.when(hearingsService.getHearingsByCaseRefNo("caseRef", "Auth", "sauth"))
+            .thenReturn(hearingsObj);
+        ResponseEntity<Object> nextHearingDetailsResponse =
+            hearingsController.getNextHearingDate("Bearer auth", "Bearer sauth", "caseRef");
+        Assertions.assertEquals(HttpStatus.OK, nextHearingDetailsResponse.getStatusCode());
+    }
+
+    @Test
+    void getNextHearingDateInternalServiceErrorTest() throws IOException, ParseException {
+        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
+        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
+
+        Mockito.when(nextHearingDetailsService.getNextHearingDate(hearings))
+            .thenThrow(new RuntimeException());
+        Mockito.when(authTokenGenerator.generate()).thenReturn("MOCK_S2S_TOKEN");
+
+        ResponseEntity<Object> nextHearingDetailsResponse =
+            hearingsController.getNextHearingDate("", "", "caseRef");
+
+        Assertions.assertEquals(
+            HttpStatus.INTERNAL_SERVER_ERROR, nextHearingDetailsResponse.getStatusCode());
+    }
+
+    @Test
+    void getNextHearingDateFeignExceptionTest() throws IOException, ParseException {
+        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
+        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
+
+        Mockito.when(nextHearingDetailsService.getNextHearingDate(hearings))
+            .thenThrow(feignException(HttpStatus.INTERNAL_SERVER_ERROR.value(), "Not found"));
+
+        ResponseEntity<Object> nextHearingDetailsResponse =
+            hearingsController.getNextHearingDate("auth", "sauth", "testcase");
+
+        Assertions.assertEquals(
+            HttpStatus.INTERNAL_SERVER_ERROR, nextHearingDetailsResponse.getStatusCode());
+    }
+
+    @Test
+    void getNextHearingDateUnauthorisedExceptionTest() throws IOException, ParseException {
+
+        ResponseEntity<Object> nextHearingDetails =
+            hearingsController.getNextHearingDate("", "", "caseRef");
+
+        Assertions.assertEquals(HttpStatus.UNAUTHORIZED, nextHearingDetails.getStatusCode());
+    }
+
+
+    public static FeignException feignException(int status, String message) {
+        return FeignException.errorStatus(
+            message,
+            Response.builder()
+                .status(status)
+                .request(Request.create(GET, EMPTY, Map.of(), new byte[] {}, UTF_8, null))
+                .build());
+    }
+
+}
