+++ src/test/java/uk/gov/hmcts/reform/hmc/api/controllers/HearingsControllerTest.java
@@ -17,7 +17,6 @@ import java.util.List;
 import java.util.Map;
 import org.json.simple.parser.ParseException;
 import org.junit.jupiter.api.Assertions;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
@@ -29,9 +28,6 @@ import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.web.server.ResponseStatusException;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.hmc.api.model.ccd.NextHearingDetails;
 import uk.gov.hmcts.reform.hmc.api.model.request.HearingValues;
 import uk.gov.hmcts.reform.hmc.api.model.response.CaseHearing;
 import uk.gov.hmcts.reform.hmc.api.model.response.HearingDaySchedule;
@@ -41,8 +37,6 @@ import uk.gov.hmcts.reform.hmc.api.model.response.linkdata.HearingLinkData;
 import uk.gov.hmcts.reform.hmc.api.services.HearingsDataService;
 import uk.gov.hmcts.reform.hmc.api.services.HearingsService;
 import uk.gov.hmcts.reform.hmc.api.services.IdamAuthService;
-import uk.gov.hmcts.reform.hmc.api.services.NextHearingDetailsService;
-import uk.gov.hmcts.reform.hmc.api.services.RoleAssignmentService;
 
 @RunWith(MockitoJUnitRunner.class)
 @ActiveProfiles("test")
@@ -52,17 +46,10 @@ class HearingsControllerTest {
 
     @Mock private IdamAuthService idamAuthService;
 
-    @Mock private AuthTokenGenerator authTokenGenerator;
-
     @Mock private HearingsDataService hearingsDataService;
 
     @Mock private HearingsService hearingsService;
 
-    @Mock private NextHearingDetailsService nextHearingDetailsService;
-
-    @Mock private RoleAssignmentService roleAssignmentService;
-
-    private Hearings hearings;
 
     @BeforeEach
     void setUp() {
@@ -319,118 +306,6 @@ class HearingsControllerTest {
     }
 
     @Test
-    void hearingsControllerNextHearingDateTest() throws IOException, ParseException {
-        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
-        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
-        ResponseEntity<Object> nextHearingDetailsResponse =
-                hearingsController.updateNextHearingDetails(
-                        "Bearer auth", "Bearer sauth", "caseRef");
-        Assertions.assertEquals(HttpStatus.OK, nextHearingDetailsResponse.getStatusCode());
-    }
-
-    @Test
-    void nextHearingDateByCaseRefNoControllerUnauthorisedExceptionTest()
-            throws IOException, ParseException {
-
-        ResponseEntity<Object> nextHearingDetails =
-                hearingsController.updateNextHearingDetails("", "", "caseRef");
-
-        Assertions.assertEquals(HttpStatus.UNAUTHORIZED, nextHearingDetails.getStatusCode());
-    }
-
-    @Test
-    void nextHearingDateByCaseRefNoControllerFeignExceptionTest()
-            throws IOException, ParseException {
-        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
-        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
-
-        Mockito.when(nextHearingDetailsService.updateNextHearingDetails("auth", hearings))
-                .thenThrow(feignException(HttpStatus.INTERNAL_SERVER_ERROR.value(), "Not found"));
-
-        ResponseEntity<Object> nextHearingUpdateResp =
-                hearingsController.updateNextHearingDetails("auth", "sauth", "testcase");
-
-        Assertions.assertEquals(
-                HttpStatus.INTERNAL_SERVER_ERROR, nextHearingUpdateResp.getStatusCode());
-    }
-
-
-    @Test
-    void nextHearingDateByCaseRefNoInternalServerErrorTest()
-        throws IOException, ParseException {
-        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
-        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
-
-        Mockito.when(nextHearingDetailsService.updateNextHearingDetails("auth", hearings))
-            .thenThrow(new RuntimeException());
-
-        ResponseEntity<Object> nextHearingUpdateResp =
-            hearingsController.updateNextHearingDetails("auth", "sauth", "testcase");
-
-        Assertions.assertEquals(
-            HttpStatus.INTERNAL_SERVER_ERROR, nextHearingUpdateResp.getStatusCode());
-    }
-
-
-    @Test
-    void getNextHearingDateTest() throws IOException, ParseException {
-        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
-        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
-        LocalDateTime testNextHearingDate = LocalDateTime.of(2024, 04, 28, 1, 0);
-        NextHearingDetails nextHearingDetails =
-                NextHearingDetails.builder()
-                        .hearingID(123L)
-                        .hearingDateTime(testNextHearingDate)
-                        .build();
-        Mockito.when(nextHearingDetailsService.getNextHearingDate(hearings))
-                .thenReturn(nextHearingDetails);
-        Mockito.when(authTokenGenerator.generate()).thenReturn("MOCK_S2S_TOKEN");
-        ResponseEntity<Object> nextHearingDetailsResponse =
-                hearingsController.getNextHearingDate("Bearer auth", "Bearer sauth", "caseRef");
-        Assertions.assertEquals(HttpStatus.OK, nextHearingDetailsResponse.getStatusCode());
-    }
-
-    @Test
-    void getNextHearingDateInternalServiceErrorTest() throws IOException, ParseException {
-        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
-        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
-
-        Mockito.when(nextHearingDetailsService.getNextHearingDate(hearings))
-                .thenThrow(new RuntimeException());
-        Mockito.when(authTokenGenerator.generate()).thenReturn("MOCK_S2S_TOKEN");
-
-        ResponseEntity<Object> nextHearingDetailsResponse =
-                hearingsController.getNextHearingDate("", "", "caseRef");
-
-        Assertions.assertEquals(
-                HttpStatus.INTERNAL_SERVER_ERROR, nextHearingDetailsResponse.getStatusCode());
-    }
-
-    @Test
-    void getNextHearingDateFeignExceptionTest() throws IOException, ParseException {
-        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
-        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
-
-        Mockito.when(nextHearingDetailsService.getNextHearingDate(hearings))
-                .thenThrow(feignException(HttpStatus.INTERNAL_SERVER_ERROR.value(), "Not found"));
-
-        ResponseEntity<Object> nextHearingDetailsResponse =
-                hearingsController.getNextHearingDate("auth", "sauth", "testcase");
-
-        Assertions.assertEquals(
-                HttpStatus.INTERNAL_SERVER_ERROR, nextHearingDetailsResponse.getStatusCode());
-    }
-
-    @Test
-    void getNextHearingDateUnauthorisedExceptionTest() throws IOException, ParseException {
-
-        ResponseEntity<Object> nextHearingDetails =
-                hearingsController.getNextHearingDate("", "", "caseRef");
-
-        Assertions.assertEquals(HttpStatus.UNAUTHORIZED, nextHearingDetails.getStatusCode());
-    }
-
-    @Test
     void allFutureHearingsByCaseRefNoControllerTest() throws IOException, ParseException {
 
         Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
@@ -557,25 +432,5 @@ class HearingsControllerTest {
 
         Assertions.assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, hearingsData1.getStatusCode());
     }
-
-    @Test
-    void assignRoleTest() {
-        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.TRUE);
-        Mockito.when(roleAssignmentService.assignHearingRoleToSysUser())
-            .thenReturn(new ResponseEntity<>(any(), HttpStatus.OK));
-        ResponseEntity<Object> response = hearingsController.assignRole("");
-        Assertions.assertEquals(HttpStatus.OK, response.getStatusCode());
-
-    }
-
-    @Test
-    void assignRoleThrowUnauthorizedExceptionTest() {
-        Mockito.when(idamAuthService.authoriseUser(any())).thenReturn(Boolean.FALSE);
-        Throwable exception = assertThrows(ResponseStatusException.class, () -> hearingsController.assignRole("Auth"));
-        Assertions.assertEquals("401 UNAUTHORIZED", exception.getMessage());
-
-    }
-
-
 
 }
